Split classes to own files
